Added support for reading and effectively changing the Displayed Name and Displayed Description (as in used in the Installer).
Prevented two Image Packages in the same WIM from having the same Name. I wanted to try to change this to have a different DISPLAYNAME value, but imagex doesn't work like that. Or at least, it doesn't do a reliable job of working like that.
Added info about sorting by Architecture to the Image Packages section of the Help file.
Changed the way the Sort by Name for Image Packages sorts to ignore architectures because of the Architecture column.
Changed -bit to -Bit in all instances of 32-bit and 64-bit.
Added finalized help files.
Fixed the Output pop-out messing with the size stored in the settings to prevent forced closures from messing up the expected size of the program on next startup.
Changed comparisons of drivers to use the Published Name instead of the identity field.
Show parsing progress no matter what, because even one driver can take long enough for it to look unresponsive.
Fixed a few tooltip display issues.
Fixed detection of edition when Sorting by Name to use the Edition value rather than the Description.
Added forced resizing code in a few key places.
Added the potential to hide output from the Output Console for certain commands. It's not currently in use.